requirements.txt
torch==1.8.0
torchvision==0.9.0
timm==0.4.9
albumentations==0.4.6
streamlit==0.78.0
pytorch_lightning==1.2.2
pandas==1.2.3
numpy==1.20.1
opencv-python==4.5.1.48
googledrivedownloader==0.4
bokeh==2.0.0
pillow==8.1.2
matplotlib==3.3.4